Extended methods moved
